requirements.txt
altgraph==0.17.4
importlib_metadata==8.5.0
packaging==25.0
pefile==2023.2.7
phidget22==1.23.20251006
PyQt5_sip==12.15.0
pywin32-ctypes==0.2.3
zipp==3.20.2
